+++ packages/framework/src/Contracts/FrontMatter/BlogPostSchema.php
@@ -3,11 +3,9 @@
 namespace Hyde\Framework\Contracts\FrontMatter;
 
 /**
- * The front matter properties supported by the following HydePHP page types and their children.
- *
  * @see \Hyde\Framework\Models\Pages\MarkdownPost
  */
-interface BlogPostSchema
+interface BlogPostSchema extends Support\FeaturedImageSchema
 {
     public const MARKDOWN_POST_SCHEMA = [
         'title'        => 'string',
@@ -22,17 +20,5 @@ interface BlogPostSchema
         'name'      => 'string',
         'username'  => 'string',
         'website'   => 'string|url',
-    ];
-
-    public const FEATURED_IMAGE_SCHEMA = [
-        'path'         => 'string',
-        'uri'          => 'string',
-        'description'  => 'string',
-        'title'        => 'string',
-        'copyright'    => 'string',
-        'license'      => 'string',
-        'licenseUrl'   => 'string',
-        'author'       => 'string',
-        'credit'       => 'string',
     ];
 }