+++ packages/framework/src/Models/Pages/DocumentationPage.php
@@ -5,7 +5,6 @@ namespace Hyde\Framework\Models\Pages;
 use Hyde\Framework\Actions\GeneratesSidebarTableOfContents;
 use Hyde\Framework\Concerns\BaseMarkdownPage;
 use Hyde\Framework\Contracts\FrontMatter\DocumentationPageSchema;
-use Hyde\Framework\Contracts\RouteContract;
 use Hyde\Framework\Models\FrontMatter;
 use Hyde\Framework\Models\Markdown;
 use Hyde\Framework\Models\Route;
@@ -15,12 +14,6 @@ class DocumentationPage extends BaseMarkdownPage implements DocumentationPageSch
     public static string $sourceDirectory = '_docs';
     public static string $outputDirectory = 'docs';
     public static string $template = 'hyde::layouts/docs';
-
-    /**
-     * @deprecated Will be replaced by navigation.group
-     * The sidebar category group, if any. Can be overridden in front matter, or by putting the source file in a subdirectory of the same category name.
-     */
-    public ?string $category = null;
 
     /** @inheritDoc */
     public function __construct(string $identifier = '', ?FrontMatter $matter = null, ?Markdown $markdown = null)
@@ -44,7 +37,7 @@ class DocumentationPage extends BaseMarkdownPage implements DocumentationPageSch
         return trim(config('docs.source_file_location_base'), '/').'/'.$this->identifier.'.md';
     }
 
-    public static function home(): ?RouteContract
+    public static function home(): ?Route
     {
         return Route::exists(static::$outputDirectory.'/index') ? Route::get(static::$outputDirectory.'/index') : null;
     }