+++ packages/realtime-compiler/src/Concerns/InteractsWithLaravel.php
@@ -16,14 +16,7 @@ trait InteractsWithLaravel
 
     protected function createApplication(): void
     {
-        // The core bootstrapping file was moved in hyde/framework v0.35.x.
-        // The old file was removed in hyde/framework v0.40.x
-
-        // To preserve backwards compatibility, we will continue to load
-        // the old bootstrap file for several minor versions.
-
-        // Temporarily add compatability for old bootstrap file location making the transition easier
-        $this->laravel = require_once file_exists(BASE_PATH.'/app/bootstrap.php') ? BASE_PATH.'/app/bootstrap.php' : BASE_PATH.'/bootstrap/app.php';
+        $this->laravel = require_once BASE_PATH.'/app/bootstrap.php';
     }
 
     protected function bootApplication(): void