+++ tests/Browser/HighLevelViewTest.php
@@ -162,8 +162,8 @@ date: 2022-01-01 12:00
 
     public function test_documentation_site_with_grouped_pages()
     {
-        $this->makeDocumentationTestPage('Page1', ['category' => 'Group 1'], true);
-        $this->makeDocumentationTestPage('Page2', ['category' => 'Group 1']);
+        $this->makeDocumentationTestPage('Page1', ['navigation.group' => 'Group 1'], true);
+        $this->makeDocumentationTestPage('Page2', ['navigation.group' => 'Group 1']);
         $this->makeDocumentationTestPage('Page3');
 
         if (! is_dir(Browser::$storeSourceAt.'/docs')) {
@@ -176,10 +176,10 @@ date: 2022-01-01 12:00
                 ->assertSee('Page1')
                 ->assertSee('Page2')
                 ->assertSee('Page3')
-                ->assertAttributeContains('#sidebar-navigation > li', 'class', 'sidebar-category')
-                ->assertSeeIn('#sidebar-navigation > li:nth-child(1) > h4.sidebar-category-heading', 'Group 1')
+                ->assertAttributeContains('#sidebar-navigation > li', 'class', 'sidebar-group')
+                ->assertSeeIn('#sidebar-navigation > li:nth-child(1) > h4.sidebar-group-heading', 'Group 1')
                 ->assertAriaAttribute('#sidebar-navigation > li:nth-child(1) > ul > li.sidebar-navigation-item.active > a', 'current', 'true')
-                ->assertSeeIn('#sidebar-navigation > li:nth-child(2) > h4.sidebar-category-heading', 'Other')
+                ->assertSeeIn('#sidebar-navigation > li:nth-child(2) > h4.sidebar-group-heading', 'Other')
                 ->screenshot('docs/with_grouped_sidebar_pages')
                 ->storeSourceAsHtml('docs/with_grouped_sidebar_pages');
         });